+++ packages/react-craft-ai-operations-history/src/components/table.jsx
@@ -4,6 +4,7 @@ import styled from 'react-emotion';
 
 export const CELL_WIDTH = 120;
 export const TIMESTAMP_CELL_WIDTH = 280;
+export const AGENT_NAME_CELL_WIDTH = 140;
 
 export function computeCellWidth(width) {
   const computedWidth = width * 15;
@@ -58,9 +59,6 @@ const Table = styled('table')`
     }
 
     padding: 10px;
-    &:last-child {
-      padding-right: 17px; // Take into account the scrollbar
-    }
 
     white-space: nowrap;
     overflow: hidden;